Better commenting + a bit of extra conditions and triggers
I'd rather not do as big of a rework as I originally planned. Sometimes I do want a modifier to persist over to different planet types in ways that don't align to the dry / wet / cold axis. The current method is clunky but works better for this. It will need good commenting and maintainance though.